[NFC] Make `DeclGroupSyntax.memberBlock` an IUO
The special decls nested inside `@abi` attributes don’t have bodies, so it’s necessary to allow `DeclGroupSyntax.memberBlock` (and the nodes that conform to it, which include the nodes for extensions and nominal types) to be `nil`.

The plan is that `nil`s will *only* occur in new syntax, not elsewhere in Swift code—even if a DeclGroupSyntax node is written without a body, the parser will still create a MemberBlockSyntax for it with missing tokens.

Since `nil` will never occur in pre-existing nodes or code, I’ve chosen to make `memberBlock` an implicitly unwrapped optional. Alternatives would be to make it a normal optional (breaking source compatibility) or to make `memberBlock` a backwards compatibility shim for an optional property with a new name.

I have nevertheless chosen to write all use sites for `memberBlock` as though the property was a normal Optional, except for ones in tests.