Introduce package manifest refactoring action "Add Package Dependency"
Start porting the package manifest editing operations from the Swift
Package Manager package over here, so that all of the syntactic
refactorings are together in one common place. These refactorings are
needed by a number of tools, including SwiftPM, SourceKit-LSP, and
(soon) the Swift compiler itself, which can all depend on
swift-syntax.

The implementation here stubs out the various types used to describe
package syntax, using simple string-backed types in place of some of
the semantic types that are part of SwiftPM itself, such as
SemanticVersion or SourceControlURL. I've also introduced the notion
of a ManifestEditRefactoringProvider to generalize over all of the
package manifest editing operations.

This commit ports over the "Add Package Dependency" command and its
tests.