Add an optional diagnostic category to diagnostics
A diagnostic category provides a category name that is used to identify
a set of related diagnostics. It can also include a documentation path
to provide more information about those diagnostics, to help guide the
user in resolving them.

Always render a category as [#CategoryName] at the end of the
diagnostic message. When we are producing colored output and there is a
documentation path, make the category name a hyperlink to the
documentation using the OSC 8 scheme.