[Diagnostic formatting] Add a category footnote printer
Categories associated with diagnostics are printed in a Markdown link style
as part of diagnostics (e.g., `[#StrictMemorySafety]`). At the end of
compilation, a tool may wish to provide documentation links to
documentation for all of the categories that showed up in that
compilation. This new function DiagnosticFormatter.categoryFootnotes()
prints the categories it is given (also as Markdown) for that purpose, e.g.,

    [#deprecated]: <http://example.com/deprecated>
    [#StrictMemorySafety]: <http://example.com/memory-safety>