Introduce FixIt.Change.textualReplacement
Introduce a new case to FixIt.Change to express an unstructured edit,
which replaces some range of source text (in a given file) with some
other source text. This is needed for some edits that aren't easily
mapped to the syntax tree, or when coming from other tools (such as
the compiler) that don't express these fixes in terms of syntax in the
first place.