[Diagnostics] Make the node and position of Diagnostic and Note optional
Some tools that wish to make use of the diagnostics machinery don't
have source locations, or even a Swift source file to point into. For
example, the compiler and driver both have diagnostics that don't
involve anything in source code, e.g., because they are diagnosing
problems at the module level such as conflicting or incorrect compiler
options. At present, the driver uses a different scheme for emitting
diagnostics, and the compiler falls back to using the LLVM diagnostics
machinery when there is no source location information.

Making the "node" and "position" of a Diagnostic optional make it
possible to express diagnostics that aren't associated with sources.
Do so, and update the diagnostic formatter to support displaying
diagnostics with no source location using the same `<unknown>:0:0:`
scheme the compiler users.