Add a new API DiagnosticsFormatter.formattedMessage(_:) for messages without source location info
Some tools want to use the diagnostics formatter to produce diagnostics
that don't relate to source code, or for which the source code isn't
available. This API allows them to do so while maintaining consistent
presentation. The first client is the compiler itself, when it doesn't
have a source location for a specific diagnostic.