[Macros] Introduce a utility to infer nonisolated conformances for macro-expanded code
The combination of isolated conformance inference with
global-actor-isolated types has introduced a source compatibility issue
with existing macros that produce conformances that are meant to be
nonisolated. Introduce a syntactic transform that identifies the
presence of nonisolated members and makes the conformances introduced
by the enclosing group "nonisolated". For example, an extension macro
that expands to:

    extension C: P {
      nonisolated func f() { }
    }

will have the macro expansion adjusted to:

    extension C: nonisolated P {
      nonisolated func f() { }
    }