[SwiftParser] SE-0478: Implement `using` declaration under an experimental feature
This is an alternative spelling of `typealias DefaultIsolation = (MainActor | nonisolated)`
that was proposed as part of SE-0478.

`using` declaration accepts an attribute or a modifier and currently
could be used to set a default actor isolation per file, but could
be extended to support other use-cases in the future.

Implementation uses `DefaultIsolationPerFile` experimental feature
flag to hide the syntax.

(cherry picked from commit 95bebf62c5498f49eead0bfaf64ddd020b9013f6)