Handle macro expansion args in placeholder expansion
Update `ExpandEditorPlaceholdersToLiteralClosures` and
`CallToTrailingClosures` such that they can handle macro expansion
exprs and decls. This unfortunately requires changing them such that
they take a `Syntax` input and output to satisfy the conformance to
`SyntaxRefactoringProvider`, but that seems preferable to making the
refactoring types themselves generic.

If in the future we decide to expose `CallLikeSyntax` as a public
protocol, we could decide to expose additional generic `refactor`
overloads here. However it's not clear there's enough clients of
these APIs to motivate doing that in this PR.