Diagnose ObjC-style string literals at top level
Both Swift parsers have always recognized `@"foo"` as an attempt to write a string literal by a programmer with Objective-C muscle memory, but some of SwiftParser’s lookahead helpers didn’t recognize that token sequence; instead they interpreted the `@` as the beginning of an attribute and therefore a missing declaration. Correct them.