[SwiftRefactor] PackageManifest: Make a couple of properties externally accessible
In it's current state `AddPackageTarget` does multiple things:

1. Updates the manifest to introduce new targets and dependencies.

2. Generates one or more auxiliary files depending on the type of
target being added (i.e. a primary file for the target and sometimes
a macro specific defintions file).

We'd like to scope the refactoring action down to manifest updates only.

The callers are going to be resposble for #2 from the list. In order to
do that `PackageTarget.sanitizedName` and `PackageTarget.Dependency.name`
have to be externally accessible (under `@_spi(PackageRefactor)` still)
to form the content of the file.