[SwiftRefactor] EditRefactoringProvider: Make `textRefactor` and `refactor` requirements throwing
The result of both has been made non-optional as well. The idea
here is to be able to produce a description for every failure
instead of `nil` result, this is something that package model
refactoring actions already do.

The only error we currently have is "not applicable" but that's
just a start.