Extend swift-parser-cli with option to apply a static build configuration
Introduce the option "--build-configuration" to this test program. The
input is JSON for a static build configuration, and that configuration
will be applied to the parse tree to remove all inactive #if regions
before doing anything else with the parse command (printing
diagnostics, printing the tree, etc.).

Also add a "print" subcommand that prints back the tree after applying
whatever transformations are described above. One can use this to
implement a basic "preprocess" step that removes all inactive #if
regions from the source code.